+++ app/src/instantsearch.js
@@ -1,5 +1,4 @@
 import instantsearch from 'instantsearch.js';
-import templates from './templates.js';
 
 import addCSS from './addCSS.js';
 import removeCSS from './removeCSS.js';
@@ -46,11 +45,14 @@ class InstantSearch {
         highlightPostTag: '</span>',
         snippetEllipsisText: '...'
       },
-      searchFunction: helper => {
+      searchFunction: ({search}) => {
+        let helper = this.instantsearch.helper;
         const query = helper.state.query;
         const optionalWords = getOptionalWords(query, this.locale);
-        this.instantsearch.helper.setQueryParameter('optionalWords', optionalWords);
-        helper.search();
+        const page = helper.getPage();
+        helper.setQueryParameter('optionalWords', optionalWords);
+        helper.setPage(page);
+        search();
       }
     });
 
@@ -75,6 +77,7 @@ class InstantSearch {
     poweredBy,
     responsive,
     subdomain,
+    templates,
     translations
   }) {
     if (!enabled) return;
@@ -83,7 +86,7 @@ class InstantSearch {
 
     let searchBoxSelector;
 
-    addCSS(templates.instantsearch.css.render({color, highlightColor}));
+    addCSS(templates.instantsearch.css({color, highlightColor}));
 
     if (reuseAutocomplete) {
       addCSS('#algolia-query { display: none }');
@@ -103,15 +106,17 @@ class InstantSearch {
     if (this.$container === null) {
       throw new Error(`[Algolia] Cannot find a container with the "${selector}" selector.`);
     }
-    this.$container.innerHTML = templates.instantsearch.layout.render({translations});
+    this.$container.innerHTML = templates.instantsearch.layout({translations});
 
-    this._handleResponsiveness({color, responsive, translations});
+    this._handleResponsiveness({responsive, templates});
 
     this.instantsearch.addWidget({
       getConfiguration: () => ({facets: ['locale.locale']}),
       init: ({helper}) => {
         // Filter by language
+        const page = helper.getPage();
         helper.addFacetRefinement('locale.locale', this.locale);
+        helper.setPage(page);
       }
     });
 
@@ -280,7 +285,7 @@ class InstantSearch {
     }, false);
   }
 
-  _handleResponsiveness({responsive}) {
+  _handleResponsiveness({templates, responsive}) {
     if (!responsive) return;
     const $mainStyle = addCSS(templates.instantsearch.responsiveCSS);